Reinstate hacky program body macro expansion
that the AC_PROTOTYPE macro relies on.
This is done by removing a level of quoting from the body argument to
the call to AC_LANG_PROGRAM.

In the given example,

    AC_PROTOTYPE(getpeername,
    [
     #include <sys/types.h>
     #include <sys/socket.h>
    ],
    [
     int a = 0;
     ARG2 * b = 0;
     ARG3 * c = 0;
     getpeername(a, b, c);
    ],
    ARG2, [struct sockaddr, void],
    ARG3, [socklen_t, size_t, int, unsigned int, long unsigned int])

AC_PROTOTYPE would do a s/ARG2/ARG2_VAL/ (and similar for the other tags)
and define ARG2_VAL macros to fill in the values.
One m4 macro expansion would then generate the final text to be
compiled.